requirements.txt
Django==4.2
django-allauth==64.0.0
social-auth-app-django==5.0.0
djangorestframework 
requests
jwt
channels==4.0
gunicorn==20.1
redis==4.5
